Move base classes to osu.Game.Rulesets.Difficulty namespace
